conditionning hard reset action
You need to set HARD_RESET env var to 'true' to get automatic hard_reset
on startup. It caused some unwanted overwrites...

+++ scripts/entrypoint.sh
@@ -42,7 +42,7 @@ if [ ! -d .git ]; then
 
     echo "Cloning from $REMOTE_URL"
     git clone -b $REMOTE_BRANCH $REMOTE_URL .
-    
+
   else
     echo "No remote configured, just init"
     git init
@@ -59,9 +59,11 @@ if [ -n "$REMOTE_NAME" ] && [ -n "$REMOTE_URL" ]; then
 
   echo "Fetch remote repo"
   git fetch $REMOTE_NAME
-  echo "Reset to upstream state"
-  git reset --hard $REMOTE_NAME/$REMOTE_BRANCH
-  git clean -xdf
+  if [ -n "$HARD_RESET" ] && [ $HARD_RESET = "yes" ]; then
+      echo "Reset to upstream state"
+      git reset --hard $REMOTE_NAME/$REMOTE_BRANCH
+      git clean -xdf
+  fi
 fi
 
 chown -R $FILES_OWNER_UID:$FILES_OWNER_GID .
